Tools: update ROS 2 launch scripts [skip ci]
- Move body of launch scripts into package for reuse.
- Add utilities module.
