+++ Tools/ros2/ardupilot_sitl/launch/virtual_ports.launch.py
@@ -20,73 +20,10 @@ Run with default arguments:
 
 ros2 launch ardupilot_sitl virtual_ports.launch.py
 """
-from typing import List
-
-from launch import LaunchContext
 from launch import LaunchDescription
-from launch import LaunchDescriptionEntity
-from launch.actions import DeclareLaunchArgument
-from launch.actions import ExecuteProcess
-from launch.actions import OpaqueFunction
-from launch.substitutions import LaunchConfiguration
-
-
-def launch_socat(
-    context: LaunchContext, *args, **kwargs
-) -> List[LaunchDescriptionEntity]:
-    """Launch a process to create virtual ports using `socat`."""
-    # Declare commands.
-    command = "socat"
-
-    # Retrieve launch arguments.
-    tty0 = LaunchConfiguration("tty0").perform(context)
-    tty1 = LaunchConfiguration("tty1").perform(context)
-
-    # Display launch arguments.
-    print(f"command:          {command}")
-    print(f"tty0:             {tty0}")
-    print(f"tty1:             {tty1}")
-
-    # Create action.
-    socat_process = ExecuteProcess(
-        cmd=[
-            [
-                "socat ",
-                "-d -d ",
-                f"pty,raw,echo=0,link={tty0} ",
-                f"pty,raw,echo=0,link={tty1} ",
-            ]
-        ],
-        shell=True,
-        output="both",
-        respawn=False,
-    )
-    return [socat_process]
+from ardupilot_sitl.launch import VirtualPortsLaunch
 
 
 def generate_launch_description() -> LaunchDescription:
     """Generate a launch description for creating virtual ports using `socat`."""
-    launch_arguments = generate_launch_arguments()
-
-    return LaunchDescription(
-        launch_arguments
-        + [
-            OpaqueFunction(function=launch_socat),
-        ]
-    )
-
-
-def generate_launch_arguments() -> List[DeclareLaunchArgument]:
-    """Generate a list of launch arguments."""
-    return [
-        DeclareLaunchArgument(
-            "tty0",
-            default_value="tty0",
-            description="Symbolic link name for tty0.",
-        ),
-        DeclareLaunchArgument(
-            "tty1",
-            default_value="tty1",
-            description="Symbolic link name for tty1.",
-        ),
-    ]
+    return VirtualPortsLaunch.generate_launch_description()
